refactor: extract home featured and infinite loading helpers
