+++ src/app/[locale]/(platform)/event/[slug]/_actions/cancel-order.ts
@@ -4,19 +4,22 @@ import { z } from 'zod'
 import { UserRepository } from '@/lib/db/queries/user'
 import { buildClobHmacSignature } from '@/lib/hmac'
 import { resolvePublicRuntimeEnv } from '@/lib/public-runtime-config.shared'
-import { TRADING_AUTH_REQUIRED_ERROR } from '@/lib/trading-auth/errors'
+import {
+  TRADING_AUTH_REQUIRED_ERROR,
+  TRADING_DEPOSIT_WALLET_REQUIRED_ERROR,
+  UNAUTHENTICATED_ERROR,
+} from '@/lib/trading-auth/errors'
 import { getUserTradingAuthSecrets } from '@/lib/trading-auth/server'
+import { DEFAULT_CANCEL_ORDER_ERROR_MESSAGE } from '@/lib/trading-flow-errors'
 
 const CancelOrderSchema = z.object({
   orderId: z.string().min(1, 'Order id is required.'),
 })
 
-const CANCEL_ORDER_ERROR = 'Unable to cancel this order right now. Please try again.'
-
 export async function cancelOrderAction(rawOrderId: string) {
   const user = await UserRepository.getCurrentUser({ disableCookieCache: true, minimal: true })
   if (!user) {
-    return { error: 'Unauthenticated.' }
+    return { error: UNAUTHENTICATED_ERROR }
   }
 
   const auth = await getUserTradingAuthSecrets(user.id)
@@ -24,7 +27,7 @@ export async function cancelOrderAction(rawOrderId: string) {
     return { error: TRADING_AUTH_REQUIRED_ERROR }
   }
   if (!user.deposit_wallet_address) {
-    return { error: 'Set up your Deposit Wallet before trading.' }
+    return { error: TRADING_DEPOSIT_WALLET_REQUIRED_ERROR }
   }
 
   const parsed = CancelOrderSchema.safeParse({ orderId: rawOrderId })
@@ -84,13 +87,13 @@ export async function cancelOrderAction(rawOrderId: string) {
           : null
 
       console.error('Failed to cancel order on CLOB.', message ?? `Status ${response.status}`)
-      return { error: message || CANCEL_ORDER_ERROR }
+      return { error: message || DEFAULT_CANCEL_ORDER_ERROR_MESSAGE }
     }
 
     return { error: null }
   }
   catch (error) {
     console.error('Failed to cancel order.', error)
-    return { error: CANCEL_ORDER_ERROR }
+    return { error: DEFAULT_CANCEL_ORDER_ERROR_MESSAGE }
   }
 }
