+++ src/app/[locale]/(platform)/event/[slug]/_actions/store-order.ts
@@ -9,7 +9,11 @@ import { OrderRepository } from '@/lib/db/queries/order'
 import { UserRepository } from '@/lib/db/queries/user'
 import { buildClobHmacSignature } from '@/lib/hmac'
 import { resolvePublicRuntimeEnv } from '@/lib/public-runtime-config.shared'
-import { TRADING_AUTH_REQUIRED_ERROR } from '@/lib/trading-auth/errors'
+import {
+  TRADING_AUTH_REQUIRED_ERROR,
+  TRADING_DEPOSIT_WALLET_REQUIRED_ERROR,
+  UNAUTHENTICATED_ERROR,
+} from '@/lib/trading-auth/errors'
 import { getUserTradingAuthSecrets } from '@/lib/trading-auth/server'
 import { normalizeAddress } from '@/lib/wallet'
 
@@ -302,7 +306,7 @@ async function readClobResponsePayload(response: {
 export async function storeOrderAction(payload: StoreOrderInput) {
   const user = await UserRepository.getCurrentUser({ disableCookieCache: true, minimal: true })
   if (!user) {
-    return { error: 'Unauthenticated.' }
+    return { error: UNAUTHENTICATED_ERROR }
   }
 
   const auth = await getUserTradingAuthSecrets(user.id)
@@ -310,7 +314,7 @@ export async function storeOrderAction(payload: StoreOrderInput) {
     return { error: TRADING_AUTH_REQUIRED_ERROR }
   }
   if (!user.deposit_wallet_address) {
-    return { error: 'Set up your Deposit Wallet before trading.' }
+    return { error: TRADING_DEPOSIT_WALLET_REQUIRED_ERROR }
   }
 
   const validated = StoreOrderSchema.safeParse(payload)
