+++ src/app/[locale]/(platform)/portfolio/_actions/cancel-all-orders.ts
@@ -3,10 +3,16 @@
 import { UserRepository } from '@/lib/db/queries/user'
 import { buildClobHmacSignature } from '@/lib/hmac'
 import { resolvePublicRuntimeEnv } from '@/lib/public-runtime-config.shared'
-import { TRADING_AUTH_REQUIRED_ERROR } from '@/lib/trading-auth/errors'
+import {
+  TRADING_AUTH_REQUIRED_ERROR,
+  TRADING_DEPOSIT_WALLET_REQUIRED_ERROR,
+  UNAUTHENTICATED_ERROR,
+} from '@/lib/trading-auth/errors'
 import { getUserTradingAuthSecrets } from '@/lib/trading-auth/server'
-
-const CANCEL_ALL_ORDERS_ERROR = 'Unable to cancel open orders right now. Please try again.'
+import {
+  DEFAULT_CANCEL_OPEN_ORDERS_ERROR_MESSAGE,
+  normalizeCancelOrdersResponse,
+} from '@/lib/trading-flow-errors'
 
 interface CancelAllOrdersResult {
   cancelled: string[]
@@ -14,25 +20,10 @@ interface CancelAllOrdersResult {
   error: string | null
 }
 
-function normalizeCancelResponse(payload: any) {
-  const cancelled = Array.isArray(payload?.cancelled)
-    ? payload.cancelled
-    : Array.isArray(payload?.canceled)
-      ? payload.canceled
-      : null
-  const notCanceled = payload?.notCanceled ?? payload?.not_canceled ?? null
-
-  if (!Array.isArray(cancelled) || !notCanceled || typeof notCanceled !== 'object' || Array.isArray(notCanceled)) {
-    return null
-  }
-
-  return { cancelled, notCanceled }
-}
-
 export async function cancelAllOrdersAction(): Promise<CancelAllOrdersResult> {
   const user = await UserRepository.getCurrentUser({ disableCookieCache: true, minimal: true })
   if (!user) {
-    return { cancelled: [], notCanceled: {}, error: 'Unauthenticated.' }
+    return { cancelled: [], notCanceled: {}, error: UNAUTHENTICATED_ERROR }
   }
 
   const auth = await getUserTradingAuthSecrets(user.id)
@@ -40,7 +31,7 @@ export async function cancelAllOrdersAction(): Promise<CancelAllOrdersResult> {
     return { cancelled: [], notCanceled: {}, error: TRADING_AUTH_REQUIRED_ERROR }
   }
   if (!user.deposit_wallet_address) {
-    return { cancelled: [], notCanceled: {}, error: 'Set up your Deposit Wallet before trading.' }
+    return { cancelled: [], notCanceled: {}, error: TRADING_DEPOSIT_WALLET_REQUIRED_ERROR }
   }
 
   const method = 'DELETE'
@@ -84,12 +75,12 @@ export async function cancelAllOrdersAction(): Promise<CancelAllOrdersResult> {
           : null
 
       console.error('Failed to cancel all orders on CLOB.', message ?? `Status ${response.status}`)
-      return { cancelled: [], notCanceled: {}, error: message || CANCEL_ALL_ORDERS_ERROR }
+      return { cancelled: [], notCanceled: {}, error: message || DEFAULT_CANCEL_OPEN_ORDERS_ERROR_MESSAGE }
     }
 
-    const normalized = normalizeCancelResponse(responsePayload)
+    const normalized = normalizeCancelOrdersResponse(responsePayload)
     if (!normalized) {
-      return { cancelled: [], notCanceled: {}, error: CANCEL_ALL_ORDERS_ERROR }
+      return { cancelled: [], notCanceled: {}, error: DEFAULT_CANCEL_OPEN_ORDERS_ERROR_MESSAGE }
     }
 
     return {
@@ -100,6 +91,6 @@ export async function cancelAllOrdersAction(): Promise<CancelAllOrdersResult> {
   }
   catch (error) {
     console.error('Failed to cancel all orders.', error)
-    return { cancelled: [], notCanceled: {}, error: CANCEL_ALL_ORDERS_ERROR }
+    return { cancelled: [], notCanceled: {}, error: DEFAULT_CANCEL_OPEN_ORDERS_ERROR_MESSAGE }
   }
 }
