+++ src/app/[locale]/(platform)/portfolio/_components/PortfolioOpenOrdersList.tsx
@@ -14,6 +14,7 @@ import { usePortfolioOpenOrdersQuery } from '@/app/[locale]/(platform)/portfolio
 import { matchesOpenOrdersSearchQuery, resolveOpenOrdersSearchParams, sortOpenOrders } from '@/app/[locale]/(platform)/portfolio/_utils/PortfolioOpenOrdersUtils'
 import { Button } from '@/components/ui/button'
 import { useDebounce } from '@/hooks/useDebounce'
+import { useInfiniteLoadMore } from '@/hooks/useInfiniteLoadMore'
 import { useOpenOrdersCacheInvalidation } from '@/hooks/useOpenOrdersCacheInvalidation'
 import { isTradingAuthRequiredError } from '@/lib/trading-auth/errors'
 import { useUser } from '@/stores/useUser'
@@ -25,12 +26,6 @@ interface PortfolioOpenOrdersListProps {
 }
 
 type OpenTradeRequirements = ReturnType<typeof useTradingOnboarding>['openTradeRequirements']
-
-interface LoadMoreStateValue {
-  key: string
-  infiniteScrollError: string | null
-  isLoadingMore: boolean
-}
 
 function useOpenOrdersFilterState(userAddress: string) {
   const [searchQuery, setSearchQuery] = useState('')
@@ -56,27 +51,6 @@ function useOpenOrdersFilterState(userAddress: string) {
     apiSearchFilters,
     apiSearchKey,
     openOrdersQueryKey,
-  }
-}
-
-function useLoadMoreState(loadMoreScopeKey: string) {
-  const [loadMoreState, setLoadMoreState] = useState<LoadMoreStateValue>({
-    key: loadMoreScopeKey,
-    infiniteScrollError: null,
-    isLoadingMore: false,
-  })
-  const scopedLoadMoreState = loadMoreState.key === loadMoreScopeKey
-    ? loadMoreState
-    : {
-        key: loadMoreScopeKey,
-        infiniteScrollError: null,
-        isLoadingMore: false,
-      }
-
-  return {
-    infiniteScrollError: scopedLoadMoreState.infiniteScrollError,
-    isLoadingMore: scopedLoadMoreState.isLoadingMore,
-    setLoadMoreState,
   }
 }
 
@@ -218,78 +192,6 @@ function useCancelOpenOrder({
   return { pendingCancelIds, handleCancelOrder }
 }
 
-function useLoadMoreOpenOrders({
-  fetchNextPage,
-  loadMoreErrorMessage,
-  loadMoreScopeKey,
-  setLoadMoreState,
-}: {
-  fetchNextPage: () => Promise<unknown>
-  loadMoreErrorMessage: string
-  loadMoreScopeKey: string
-  setLoadMoreState: (value: LoadMoreStateValue) => void
-}) {
-  return useCallback(() => {
-    setLoadMoreState({
-      key: loadMoreScopeKey,
-      infiniteScrollError: null,
-      isLoadingMore: true,
-    })
-
-    fetchNextPage()
-      .then(() => {
-        setLoadMoreState({
-          key: loadMoreScopeKey,
-          infiniteScrollError: null,
-          isLoadingMore: false,
-        })
-      })
-      .catch((error: any) => {
-        setLoadMoreState({
-          key: loadMoreScopeKey,
-          infiniteScrollError: error?.name === 'AbortError' ? null : error?.message || loadMoreErrorMessage,
-          isLoadingMore: false,
-        })
-      })
-  }, [fetchNextPage, loadMoreErrorMessage, loadMoreScopeKey, setLoadMoreState])
-}
-
-function useInfiniteScrollSentinel({
-  hasNextPage,
-  infiniteScrollError,
-  isFetchingNextPage,
-  isLoadingMore,
-  loadMoreOpenOrders,
-}: {
-  hasNextPage: boolean
-  infiniteScrollError: string | null
-  isFetchingNextPage: boolean
-  isLoadingMore: boolean
-  loadMoreOpenOrders: () => void
-}): { loadMoreRef: RefObject<HTMLDivElement | null> } {
-  const loadMoreRef = useRef<HTMLDivElement | null>(null)
-
-  useEffect(function observeLoadMoreSentinel() {
-    if (!hasNextPage || !loadMoreRef.current) {
-      return undefined
-    }
-
-    const observer = new IntersectionObserver((entries) => {
-      const [entry] = entries
-      if (entry?.isIntersecting && !isFetchingNextPage && !isLoadingMore && !infiniteScrollError) {
-        loadMoreOpenOrders()
-      }
-    }, { rootMargin: '200px' })
-
-    observer.observe(loadMoreRef.current)
-    return function disconnectLoadMoreObserver() {
-      observer.disconnect()
-    }
-  }, [hasNextPage, infiniteScrollError, isFetchingNextPage, isLoadingMore, loadMoreOpenOrders])
-
-  return { loadMoreRef }
-}
-
 function promptTradingAuthForOpenOrdersError({
   error,
   hasPromptedTradingAuthRef,
@@ -330,7 +232,6 @@ export default function PortfolioOpenOrdersList({ userAddress }: PortfolioOpenOr
     openOrdersQueryKey,
   } = useOpenOrdersFilterState(userAddress)
   const loadMoreScopeKey = `${userAddress}:${apiSearchKey}:${searchQuery}:${sortBy}`
-  const { infiniteScrollError, isLoadingMore, setLoadMoreState } = useLoadMoreState(loadMoreScopeKey)
 
   const {
     status,
@@ -390,19 +291,17 @@ export default function PortfolioOpenOrdersList({ userAddress }: PortfolioOpenOr
     openTradeRequirements,
   })
 
-  const loadMoreOpenOrders = useLoadMoreOpenOrders({
-    fetchNextPage,
-    loadMoreErrorMessage: t('Failed to load more open orders'),
-    loadMoreScopeKey,
-    setLoadMoreState,
-  })
-
-  const { loadMoreRef } = useInfiniteScrollSentinel({
-    hasNextPage,
+  const {
     infiniteScrollError,
-    isFetchingNextPage,
     isLoadingMore,
-    loadMoreOpenOrders,
+    loadMoreRef,
+    loadMore,
+  } = useInfiniteLoadMore({
+    loadMoreScopeKey,
+    hasNextPage,
+    isFetchingNextPage,
+    fetchNextPage,
+    errorMessage: t('Failed to load more open orders'),
   })
 
   const emptyText = userAddress
@@ -441,7 +340,7 @@ export default function PortfolioOpenOrdersList({ userAddress }: PortfolioOpenOr
         infiniteScrollError={infiniteScrollError}
         isLoadingMore={isLoadingMore}
         loadMoreRef={loadMoreRef}
-        onRetryLoadMore={loadMoreOpenOrders}
+        onRetryLoadMore={loadMore}
         onCancelOrder={handleCancelOrder}
         pendingCancelIds={pendingCancelIds}
       />
